refactor to make package module friendly
